fixed learning rate

+++ Practical 2/code/settings.py
@@ -43,7 +43,7 @@ def parse_settings():
     # Training
     parser.add_argument('--cuda', type=int, default=None,
                         help='Cuda device number or None.')
-    parser.add_argument('--lr', type=int, default=1e-3,
+    parser.add_argument('--lr', type=float, default=1e-3,
                         help='Learning rate.')
     parser.add_argument('--num_epochs', type=int, default=10,
                         help='Number of epochs over the data.')
